serverless/requirements.txt
boto3==1.34.23
Cerberus==1.3.5
cryptography==41.0.7
fast-ulid==0.2.0
firebase-admin==6.3.0
Flask==3.0.1
Flask-Cognito==1.20
Markdown==3.5.2
Pillow==10.1.0
PyYAML==6.0.1
python-ulid==2.2.0
qrcode==7.4.2
requests==2.31.0
Werkzeug==3.0.1
